refactor: move all tests to one folder
