+++ tests/backendTests/userAndTeams.test.js
@@ -1,9 +1,9 @@
-import { chrome } from "../__mocks__/chromeMock.js";
+import { chrome } from "../mocks/chromeMock.js";
 global.chrome = chrome;
-import _, { getDate } from "../public/userAndTeams.js";
-import { localStorageMock } from "../__mocks__/testMock.js";
-import { setDB } from "../public/firebaseInit.js";
-import { db, get, set, update } from "../__mocks__/databaseMock.js";
+import _, { getDate } from "../../public/userAndTeams.js";
+import { localStorageMock } from "../mocks/testMock.js";
+import { setDB } from "../../public/firebaseInit.js";
+import { db, get, set, update } from "../mocks/databaseMock.js";
 
 //localStorageMock.setItem("test", 2);
 //console.log(store);
