+++ tests/mocks/chromeMock.js
