+++ tests/mocks/testMock.js
